abstract paging logic
